Revert "Move transcoding constant/utils back to `pipeline`"
This reverts commit 1aa743482f9cdcfd4382128c6e1ab10bc5e6a2cd.
