+++ kedro/framework/context/context.py
@@ -15,7 +15,7 @@ from pluggy import PluginManager
 from kedro.config import AbstractConfigLoader, MissingConfigException
 from kedro.framework.project import settings
 from kedro.io import DataCatalog
-from kedro.pipeline.pipeline import _transcode_split
+from kedro.pipeline._transcoding import _transcode_split
 
 
 def _is_relative_path(path_string: str) -> bool:
