+++ kedro/pipeline/node.py
@@ -13,6 +13,8 @@ from warnings import warn
 
 from more_itertools import spy, unzip
 
+from ._transcoding import _strip_transcoding
+
 
 class Node:
     """``Node`` is an auxiliary class facilitating the operations required to
@@ -528,8 +530,6 @@ class Node:
             )
 
     def _validate_inputs_dif_than_outputs(self) -> None:
-        from kedro.pipeline.pipeline import _strip_transcoding
-
         common_in_out = set(map(_strip_transcoding, self.inputs)).intersection(
             set(map(_strip_transcoding, self.outputs))
         )
